requirements.txt
transformers
datasets==2.14.6
deepspeed==0.11.1
pytorch_lightning==1.9.1